Moved lerp function
